requirements.txt
envparse==0.1.6
future==0.15.2
oauthlib==1.0.3
python-telegram-bot==2.9.2
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
tweepy==3.4.0
wheel==0.24.0
